requirements.txt
faiss-cpu
sentence-transformers
datasets
torch
